app/requirements.txt
Flask==1.0.2
Werkzeug==0.14.1
google-auth-httplib2==0.0.3
google-api-python-client==1.7.11